+++ packages/core/src/indexing-store/historicalStore.ts
@@ -1,10 +1,10 @@
-import { StoreError } from "@/common/errors.js";
 import { HeadlessKysely } from "@/database/kysely.js";
 import type { NamespaceInfo } from "@/database/service.js";
 import type { Schema } from "@/schema/types.js";
 import { sql } from "kysely";
 import type { Row, WhereInput, WriteIndexingStore } from "./store.js";
 import { decodeRow, encodeRow, encodeValue } from "./utils/encoding.js";
+import { parseStoreError } from "./utils/errors.js";
 import { buildWhereConditions } from "./utils/filter.js";
 
 const MAX_BATCH_SIZE = 1_000 as const;
@@ -34,26 +34,17 @@ export const getHistoricalIndexingStore = ({
     return db.wrap({ method: `${tableName}.create` }, async () => {
       const createRow = encodeRow({ id, ...data }, table, kind);
 
-      try {
-        const row = await db
-          .withSchema(namespaceInfo.userNamespace)
-          .insertInto(tableName)
-          .values(createRow)
-          .returningAll()
-          .executeTakeFirstOrThrow();
+      const row = await db
+        .withSchema(namespaceInfo.userNamespace)
+        .insertInto(tableName)
+        .values(createRow)
+        .returningAll()
+        .executeTakeFirstOrThrow()
+        .catch((err) => {
+          throw parseStoreError(err, { id, ...data });
+        });
 
-        return decodeRow(row, table, kind);
-      } catch (err) {
-        const error = err as Error;
-        throw (kind === "sqlite" &&
-          error.message.includes("UNIQUE constraint failed")) ||
-          (kind === "postgres" &&
-            error.message.includes("violates unique constraint"))
-          ? new StoreError(
-              `Cannot create ${tableName} record with ID ${id} because a record already exists with that ID (UNIQUE constraint violation). Hint: Did you forget to await the promise returned by a store method? Or, consider using ${tableName}.upsert().`,
-            )
-          : error;
-      }
+      return decodeRow(row, table, kind);
     });
   },
   createMany: async ({
@@ -69,30 +60,21 @@ export const getHistoricalIndexingStore = ({
 
     for (let i = 0, len = data.length; i < len; i += MAX_BATCH_SIZE) {
       await db.wrap({ method: `${tableName}.createMany` }, async () => {
-        try {
-          const createRows = data
-            .slice(i, i + MAX_BATCH_SIZE)
-            .map((d) => encodeRow(d, table, kind));
+        const createRows = data
+          .slice(i, i + MAX_BATCH_SIZE)
+          .map((d) => encodeRow(d, table, kind));
 
-          const _rows = await db
-            .withSchema(namespaceInfo.userNamespace)
-            .insertInto(tableName)
-            .values(createRows)
-            .returningAll()
-            .execute();
+        const _rows = await db
+          .withSchema(namespaceInfo.userNamespace)
+          .insertInto(tableName)
+          .values(createRows)
+          .returningAll()
+          .execute()
+          .catch((err) => {
+            throw parseStoreError(err, data.length > 0 ? data[0] : {});
+          });
 
-          rows.push(...(_rows as Row[]));
-        } catch (err) {
-          const error = err as Error;
-          throw (kind === "sqlite" &&
-            error.message.includes("UNIQUE constraint failed")) ||
-            (kind === "postgres" &&
-              error.message.includes("violates unique constraint"))
-            ? new StoreError(
-                `Cannot createMany ${tableName} records because one or more records already exist (UNIQUE constraint violation). Hint: Did you forget to await the promise returned by a store method?`,
-              )
-            : error;
-        }
+        rows.push(...(_rows as Row[]));
       });
     }
 
@@ -114,37 +96,36 @@ export const getHistoricalIndexingStore = ({
     return db.wrap({ method: `${tableName}.update` }, async () => {
       const encodedId = encodeValue(id, table.id, kind);
 
-      let updateRow: ReturnType<typeof encodeRow>;
-
+      let updateObject: Partial<Omit<Row, "id">>;
       if (typeof data === "function") {
-        // Find the latest version of this instance.
         const latestRow = await db
           .withSchema(namespaceInfo.userNamespace)
           .selectFrom(tableName)
           .selectAll()
           .where("id", "=", encodedId)
-          .executeTakeFirst();
-        if (!latestRow)
-          throw new StoreError(
-            `Cannot update ${tableName} record with ID ${id} because no existing record was found with that ID. Consider using ${tableName}.upsert(), or create the record before updating it. Hint: Did you forget to await the promise returned by a store method?`,
-          );
+          .executeTakeFirstOrThrow()
+          .catch((err) => {
+            throw parseStoreError(err, { id, data: "(function)" });
+          });
 
-        const current = decodeRow(latestRow, table, kind);
-        const updateObject = data({ current });
-        updateRow = encodeRow({ id, ...updateObject }, table, kind);
+        updateObject = data({ current: decodeRow(latestRow, table, kind) });
       } else {
-        updateRow = encodeRow({ id, ...data }, table, kind);
+        updateObject = data;
       }
+      const updateRow = encodeRow({ id, ...updateObject }, table, kind);
+
       const row = await db
+        .withSchema(namespaceInfo.userNamespace)
         .updateTable(tableName)
         .set(updateRow)
         .where("id", "=", encodedId)
         .returningAll()
-        .executeTakeFirstOrThrow();
+        .executeTakeFirstOrThrow()
+        .catch((err) => {
+          throw parseStoreError(err, { id, ...updateObject });
+        });
 
-      const result = decodeRow(row, table, kind);
-
-      return result;
+      return decodeRow(row, table, kind);
     });
   },
   updateMany: async ({
@@ -163,28 +144,29 @@ export const getHistoricalIndexingStore = ({
     return db.wrap({ method: `${tableName}.updateMany` }, async () => {
       if (typeof data === "function") {
         const rows = await db.transaction().execute(async (tx) => {
-          let query = tx
+          const latestRows = await tx
             .withSchema(namespaceInfo.userNamespace)
             .selectFrom(tableName)
-            .selectAll();
-
-          if (where) {
-            query = query.where((eb) =>
+            .selectAll()
+            .where((eb) =>
               buildWhereConditions({
                 eb,
                 where,
                 table,
                 encoding: kind,
               }),
-            );
-          }
+            )
+            .execute();
 
-          const latestRows = await query.execute();
-          const rows: Row[] = [];
+          const rows: { [x: string]: any }[] = [];
           for (const latestRow of latestRows) {
             const current = decodeRow(latestRow, table, kind);
             const updateObject = data({ current });
-            const updateRow = encodeRow(updateObject, table, kind);
+            // Here, `latestRow` is already encoded, so we need to exclude it from `encodeRow`.
+            const updateRow = {
+              id: latestRow.id,
+              ...encodeRow(updateObject, table, kind),
+            };
 
             const row = await tx
               .withSchema(namespaceInfo.userNamespace)
@@ -192,9 +174,11 @@ export const getHistoricalIndexingStore = ({
               .set(updateRow)
               .where("id", "=", latestRow.id)
               .returningAll()
-              .executeTakeFirstOrThrow();
-
-            rows.push(row as Row);
+              .executeTakeFirstOrThrow()
+              .catch((err) => {
+                throw parseStoreError(err, updateObject);
+              });
+            rows.push(row);
           }
 
           return rows;
@@ -205,31 +189,30 @@ export const getHistoricalIndexingStore = ({
         const updateRow = encodeRow(data, table, kind);
 
         const rows = await db
-          .with("latestRows(id)", (db) => {
-            let query = db
+          .with("latestRows(id)", (db) =>
+            db
               .withSchema(namespaceInfo.userNamespace)
               .selectFrom(tableName)
-              .select("id");
-
-            if (where) {
-              query = query.where((eb) =>
+              .select("id")
+              .where((eb) =>
                 buildWhereConditions({
                   eb,
                   where,
                   table,
                   encoding: kind,
                 }),
-              );
-            }
-            return query;
-          })
+              ),
+          )
           .withSchema(namespaceInfo.userNamespace)
           .updateTable(tableName)
           .set(updateRow)
           .from("latestRows")
           .where(`${tableName}.id`, "=", sql.ref("latestRows.id"))
           .returningAll()
-          .execute();
+          .execute()
+          .catch((err) => {
+            throw parseStoreError(err, data);
+          });
 
         return rows.map((row) => decodeRow(row, table, kind));
       }
@@ -255,7 +238,6 @@ export const getHistoricalIndexingStore = ({
       const createRow = encodeRow({ id, ...create }, table, kind);
 
       if (typeof update === "function") {
-        // Find the latest version of this instance.
         const latestRow = await db
           .withSchema(namespaceInfo.userNamespace)
           .selectFrom(tableName)
@@ -263,14 +245,20 @@ export const getHistoricalIndexingStore = ({
           .where("id", "=", encodedId)
           .executeTakeFirst();
 
-        // If there is no latest version, insert a new version using the create data.
         if (latestRow === undefined) {
           const row = await db
             .withSchema(namespaceInfo.userNamespace)
             .insertInto(tableName)
             .values(createRow)
             .returningAll()
-            .executeTakeFirstOrThrow();
+            .executeTakeFirstOrThrow()
+            .catch((err) => {
+              const prettyObject: any = { id };
+              for (const [key, value] of Object.entries(create))
+                prettyObject[`create.${key}`] = value;
+              prettyObject.update = "(function)";
+              throw parseStoreError(err, prettyObject);
+            });
 
           return decodeRow(row, table, kind);
         }
@@ -285,7 +273,16 @@ export const getHistoricalIndexingStore = ({
           .set(updateRow)
           .where("id", "=", encodedId)
           .returningAll()
-          .executeTakeFirstOrThrow();
+          .executeTakeFirstOrThrow()
+          .catch((err) => {
+            const prettyObject: any = { id };
+            for (const [key, value] of Object.entries(create))
+              prettyObject[`create.${key}`] = value;
+            for (const [key, value] of Object.entries(updateObject))
+              prettyObject[`update.${key}`] = value;
+            throw parseStoreError(err, prettyObject);
+          });
+
         return decodeRow(row, table, kind);
       } else {
         const updateRow = encodeRow({ id, ...update }, table, kind);
@@ -296,7 +293,15 @@ export const getHistoricalIndexingStore = ({
           .values(createRow)
           .onConflict((oc) => oc.column("id").doUpdateSet(updateRow))
           .returningAll()
-          .executeTakeFirstOrThrow();
+          .executeTakeFirstOrThrow()
+          .catch((err) => {
+            const prettyObject: any = { id };
+            for (const [key, value] of Object.entries(create))
+              prettyObject[`create.${key}`] = value;
+            for (const [key, value] of Object.entries(update))
+              prettyObject[`update.${key}`] = value;
+            throw parseStoreError(err, prettyObject);
+          });
 
         return decodeRow(row, table, kind);
       }
@@ -319,7 +324,10 @@ export const getHistoricalIndexingStore = ({
         .deleteFrom(tableName)
         .where("id", "=", encodedId)
         .returning(["id"])
-        .executeTakeFirst();
+        .executeTakeFirst()
+        .catch((err) => {
+          throw parseStoreError(err, { id });
+        });
 
       return !!deletedRow;
     });