+++ packages/core/src/indexing/service.ts
@@ -1,5 +1,6 @@
 import type { IndexingFunctions } from "@/build/configAndIndexingFunctions.js";
 import type { Common } from "@/common/common.js";
+import { getBaseError } from "@/common/errors.js";
 import type { Network } from "@/config/networks.js";
 import {
   type EventSource,
@@ -429,10 +430,8 @@ const executeSetup = async (
   const eventName = `${event.contractName}:setup`;
   const indexingFunction = indexingFunctions[eventName];
 
-  const metricLabel = {
-    event: eventName,
-    network: networkByChainId[event.chainId].name,
-  };
+  const networkName = networkByChainId[event.chainId].name;
+  const metricLabel = { event: eventName, network: networkName };
 
   try {
     // set currentEvent
@@ -455,7 +454,7 @@ const executeSetup = async (
     );
   } catch (error_) {
     if (indexingService.isKilled) return { status: "killed" };
-    const error = error_ as Error & { meta?: string };
+    const error = getBaseError(error_);
 
     common.metrics.ponder_indexing_function_error_total.inc(metricLabel);
 
@@ -467,7 +466,7 @@ const executeSetup = async (
 
     common.logger.error({
       service: "indexing",
-      msg: `Error while processing "${eventName}" event at chainId=${decodedCheckpoint.chainId}, block=${decodedCheckpoint.blockNumber}: `,
+      msg: `Error while processing '${eventName}' event in '${networkName}' block ${decodedCheckpoint.blockNumber}`,
       error,
     });
 
@@ -496,10 +495,8 @@ const executeLog = async (
   const eventName = `${event.contractName}:${event.logEventName}`;
   const indexingFunction = indexingFunctions[eventName];
 
-  const metricLabel = {
-    event: eventName,
-    network: networkByChainId[event.chainId].name,
-  };
+  const networkName = networkByChainId[event.chainId].name;
+  const metricLabel = { event: eventName, network: networkName };
 
   try {
     // set currentEvent
@@ -530,23 +527,19 @@ const executeLog = async (
     );
   } catch (error_) {
     if (indexingService.isKilled) return { status: "killed" };
-    const error = error_ as Error & { meta?: string };
+    const error = getBaseError(error_);
 
     common.metrics.ponder_indexing_function_error_total.inc(metricLabel);
 
     const decodedCheckpoint = decodeCheckpoint(event.encodedCheckpoint);
 
-    if (error.meta) {
-      error.meta += `\nEvent args:\n${prettyPrint(event.event.args)}`;
-    } else {
-      error.meta = `Event args:\n${prettyPrint(event.event.args)}`;
-    }
+    error.meta.push(`Event arguments:\n${prettyPrint(event.event.args)}`);
 
     addUserStackTrace(error, common.options);
 
     common.logger.error({
       service: "indexing",
-      msg: `Error while processing "${eventName}" event at chainId=${decodedCheckpoint.chainId}, block=${decodedCheckpoint.blockNumber}: `,
+      msg: `Error while processing '${eventName}' event in '${networkName}' block ${decodedCheckpoint.blockNumber}`,
       error,
     });