+++ packages/core/src/sync-historical/service.ts
@@ -279,7 +279,7 @@ export class HistoricalSyncService extends Emittery<HistoricalSyncEvents> {
             service: "historical",
             msg: `Started sync with ${formatPercentage(
               Math.min(1, cachedBlockCount / (targetBlockCount || 1)),
-            )} cached (contract=${source.contractName} network=${
+            )} cached (contract=${source.contractName}, network=${
               this.network.name
             })`,
           });
@@ -447,7 +447,7 @@ export class HistoricalSyncService extends Emittery<HistoricalSyncEvents> {
             service: "historical",
             msg: `Started sync with ${formatPercentage(
               cacheRate,
-            )} cached (contract=${source.contractName} network=${
+            )} cached (contract=${source.contractName}, network=${
               this.network.name
             })`,
           });
@@ -533,7 +533,7 @@ export class HistoricalSyncService extends Emittery<HistoricalSyncEvents> {
           service: "historical",
           msg: `Sync is ${formatPercentage(progress ?? 0)} complete${
             eta !== undefined ? ` with ~${formatEta(eta)} remaining` : ""
-          } (contract=${contractName})`,
+          } (contract=${contractName}, network=${this.network.name})`,
           network: this.network.name,
         });
       });